Use OOP in script.js
Put the whole code into a new class named BooksList. It has the argument
dataSource. In constructor, create some usual properties of the class and put
called functions into it. Create new method getElements(). Replace all
dom elements in methods with those in getElements() and custom these
methods. After this outside of this class make a constant app which is a
new instance of the class BooksList.